backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.7.0
pymongo==4.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
pydantic[email-validator]==2.10.4
python-dotenv==1.0.1
langchain>=1.0.0
langchain-core>=0.3.0
langchain-community>=0.3.14
langchain-ollama>=0.3.0
langchain-chroma>=0.2.0
langgraph>=0.4.0
langgraph-checkpoint>=2.0.0
chromadb>=0.5.23
rank-bm25>=0.2.2
python-socketio==5.12.1
httpx==0.28.1
pandas==2.2.3
numpy==2.2.1
